+++ src/aws/src/lambda.ts
@@ -8,9 +8,9 @@ import type {
   Response,
 } from '@heliosjs/core/types';
 import { CONTROLLER_REQUEST } from '@heliosjs/core/constants';
-import { ApplicationError, getErrorType } from '@heliosjs/core/utils';
+import { ApplicationError, getErrorType, setRolesExtractor } from '@heliosjs/core/utils';
 import type { Context, Handler } from 'aws-lambda';
-import type { ILambdaAdapter, LambdaEvent, Plugin as LambdaPlugin } from './types/aws';
+import type { ILambdaAdapter, LambdaEvent, LambdaOptions, Plugin as LambdaPlugin } from './types/aws';
 import { getEventType, Plugin, RequestFactory, ResponseFactory } from './utils/aws';
 
 /**
@@ -34,9 +34,11 @@ export class Helios extends Plugin implements ILambdaAdapter {
    *
    * @param controller - Decorated controller class that handles incoming requests.
    */
-  constructor(controller: ControllerClass) {
+  constructor(controller: ControllerClass, options?: LambdaOptions) {
     super();
-
+    if (options?.rbac?.getRoles) {
+      setRolesExtractor(options.rbac.getRoles);
+    }
     this.controller = this.compileController(controller);
     this.handler = this.createHandler();
   }