+++ src/middlewares/src/index.ts
@@ -5,6 +5,7 @@ export * from './cors';
 export * from './guard';
 export * from './interceptor';
 export * from './pipe';
+export * from './roles';
 export * from './sanitize';
 export * from './status';
 export * from './use';